Refactor testing fixtures for cleaner separation
